Separate few of the logic in other files
