+++ Rocket.Chat/Controllers/Chat/ChatControllerMessageCellProtocol.swift
@@ -1,0 +1,42 @@
+//
+//  ChatControllerMessageCellProtocol.swift
+//  Rocket.Chat
+//
+//  Created by Rafael Kellermann Streit on 17/12/16.
+//  Copyright © 2016 Rocket.Chat. All rights reserved.
+//
+
+import UIKit
+
+
+extension ChatViewController: ChatMessageCellProtocol {
+    
+    func openURL(url: URL) {
+        let controller = SFSafariViewController(url: url)
+        present(controller, animated: true, completion: nil)
+    }
+    
+    func openURLFromCell(url: MessageURL) {
+        guard let targetURL = url.targetURL else { return }
+        guard let destinyURL = URL(string: targetURL) else { return }
+        let controller = SFSafariViewController(url: destinyURL)
+        present(controller, animated: true, completion: nil)
+    }
+    
+    func openVideoFromCell(attachment: Attachment) {
+        guard let videoURL = attachment.fullVideoURL() else { return }
+        let controller = MobilePlayerViewController(contentURL: videoURL)
+        controller.title = attachment.title
+        controller.activityItems = [attachment.title, videoURL]
+        present(controller, animated: true, completion: nil)
+    }
+    
+    func openImageFromCell(attachment: Attachment, thumbnail: UIImageView) {
+        if let image = thumbnail.image {
+            mediaFocusViewController.show(image, from: thumbnail)
+        } else {
+            mediaFocusViewController.showImage(from: attachment.fullImageURL(), from: thumbnail)
+        }
+    }
+    
+}
