+++ Rocket.Chat/Controllers/Chat/ChatViewController.swift
@@ -165,82 +165,6 @@ class ChatViewController: SLKTextViewController {
         scrollToBottom()
     }
     
-    override func didChangeAutoCompletionPrefix(_ prefix: String, andWord word: String) {
-        searchResult = [:]
-        
-        if prefix == "@" && word.characters.count > 0 {
-            guard let users = try? Realm().objects(User.self).filter(NSPredicate(format: "username BEGINSWITH[c] %@", word)) else { return }
-
-            for user in users {
-                if let username = user.username {
-                    searchResult[username] = user
-                }
-            }
-            
-            if "here".contains(word) {
-                searchResult["here"] = UIImage(named: "Hashtag")
-            }
-            
-            if "all".contains(word) {
-                searchResult["all"] = UIImage(named: "Hashtag")
-            }
-
-        } else if prefix == "#" && word.characters.count > 0 {
-            guard let channels = try? Realm().objects(Subscription.self).filter("auth != nil && (privateType == 'c' || privateType == 'p') && name BEGINSWITH[c] %@", word) else { return }
-            
-            for channel in channels {
-                searchResult[channel.name] = channel.type == .channel ? UIImage(named: "Hashtag") : UIImage(named: "Lock")
-            }
-
-        }
-        
-        let show = (searchResult.count > 0)
-        showAutoCompletionView(show)
-    }
-    
-    override func heightForAutoCompletionView() -> CGFloat {
-        return AutocompleteCell.minimumHeight * CGFloat(searchResult.count)
-    }
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return searchResult.keys.count
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        return autoCompletionCellForRowAtIndexPath(indexPath)
-    }
-    
-    override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return AutocompleteCell.minimumHeight
-    }
-    
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let key = Array(searchResult.keys)[indexPath.row]
-        acceptAutoCompletion(with: "\(key) ", keepPrefix: true)
-    }
-    
-    func autoCompletionCellForRowAtIndexPath(_ indexPath: IndexPath) -> AutocompleteCell {
-        let key = Array(searchResult.keys)[indexPath.row]
-        let cell = autoCompletionView.dequeueReusableCell(withIdentifier: AutocompleteCell.identifier) as! AutocompleteCell
-        cell.selectionStyle = .default
-        
-        if let user = searchResult[key] as? User {
-            cell.avatarView.isHidden = false
-            cell.imageViewIcon.isHidden = true
-            cell.avatarView.user = user
-        } else {
-            cell.avatarView.isHidden = true
-            cell.imageViewIcon.isHidden = false
-            
-            if let image = searchResult[key] as? UIImage {
-                cell.imageViewIcon.image = image.imageWithTint(.lightGray)
-            }
-        }
-
-        cell.labelTitle.text = key
-        return cell
-    }
-    
     
     // MARK: Message
     
@@ -508,59 +432,3 @@ extension ChatViewController: ChatPreviewModeViewProtocol {
     
 }
 
-
-// MARK: ChatURLCellProtocol
-
-extension ChatViewController: ChatMessageCellProtocol {
-    
-    func openURL(url: URL) {
-        let controller = SFSafariViewController(url: url)
-        present(controller, animated: true, completion: nil)
-    }
-    
-    func openURLFromCell(url: MessageURL) {
-        guard let targetURL = url.targetURL else { return }
-        guard let destinyURL = URL(string: targetURL) else { return }
-        let controller = SFSafariViewController(url: destinyURL)
-        present(controller, animated: true, completion: nil)
-    }
-    
-    func openVideoFromCell(attachment: Attachment) {
-        guard let videoURL = attachment.fullVideoURL() else { return }
-        let controller = MobilePlayerViewController(contentURL: videoURL)
-        controller.title = attachment.title
-        controller.activityItems = [attachment.title, videoURL]
-        present(controller, animated: true, completion: nil)
-    }
-    
-    func openImageFromCell(attachment: Attachment, thumbnail: UIImageView) {
-        if let image = thumbnail.image {
-            mediaFocusViewController.show(image, from: thumbnail)
-        } else {
-            mediaFocusViewController.showImage(from: attachment.fullImageURL(), from: thumbnail)
-        }
-    }
-
-}
-
-
-// MARK: SocketConnectionHandler
-
-extension ChatViewController: SocketConnectionHandler {
-    
-    func socketDidConnect(socket: SocketManager) {
-        chatHeaderViewOffline?.removeFromSuperview()
-        updateSubscriptionMessages()
-        rightButton.isEnabled = true
-    }
-    
-    func socketDidDisconnect(socket: SocketManager) {
-        chatHeaderViewOffline?.removeFromSuperview()
-        
-        let headerView = ChatHeaderViewOffline.instanceFromNib() as! ChatHeaderViewOffline
-        headerView.frame = CGRect(x: 0, y: 0, width: view.frame.width, height: headerView.frame.height)
-        view.addSubview(headerView)
-        chatHeaderViewOffline = headerView
-    }
-    
-}
